+++ src/main/java/com/github/marcindabrowski/example/roomsbooking/application/rest/BookRoomsRequest.java
@@ -2,7 +2,8 @@ package com.github.marcindabrowski.example.roomsbooking.application.rest;
 
 import com.github.marcindabrowski.example.roomsbooking.domain.model.HotelFreeRooms;
 
+import java.math.BigDecimal;
 import java.util.List;
 
-public record BookRoomsRequest(HotelFreeRooms freeRooms, List<Double> potentialGuests) {
+public record BookRoomsRequest(HotelFreeRooms freeRooms, List<BigDecimal> potentialGuests) {
 }
